Fix bug in original Flesch Reading Ease and Gunning Fog #63
The original code (ce5a5ad) was removing all question marks from the
turns and then computing the values based on this single long text,
therefore, from the point of view of the metrics the input was a single
loooong sentence, breaking the formulas since they expect to have a set
of sentences. Plus, different turns should also be considered different
sentences by default, which was not the case before.